refactor: reorganize downloads package into worker/validation/retry subdirectories
