+++ src/rheo/downloads/retry/__init__.py
@@ -1,0 +1,13 @@
+"""Retry logic - error categorisation and retry handling."""
+
+from .base import BaseRetryHandler
+from .categoriser import ErrorCategoriser
+from .handler import RetryHandler
+from .null import NullRetryHandler
+
+__all__ = [
+    "BaseRetryHandler",
+    "RetryHandler",
+    "NullRetryHandler",
+    "ErrorCategoriser",
+]
