+++ src/rheo/downloads/validation/__init__.py
@@ -1,0 +1,11 @@
+"""File validation - hash checking and integrity verification."""
+
+from .base import BaseFileValidator
+from .null import NullFileValidator
+from .validator import FileValidator
+
+__all__ = [
+    "BaseFileValidator",
+    "FileValidator",
+    "NullFileValidator",
+]
