+++ src/rheo/downloads/worker/__init__.py
@@ -1,0 +1,5 @@
+"""Download worker implementations."""
+
+from .worker import DownloadWorker
+
+__all__ = ["DownloadWorker"]
